Debug File Java Core
